ai_agent/requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.32.5
urllib3==1.26.18
python-dotenv==1.0.0
click==8.1.7
itsdangerous==2.1.2
jinja2==3.1.2
markupsafe==2.1.3
typing-extensions==4.14.1
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7